requirements.txt
numpy
pandas
scikit-learn
openml
matplotlib
nltk
tqdm
sentence-transformers
gensim
openai
google-generativeai
ollama
langdetect
requests
seaborn
pulp
pyswarm
spacy
